Links are generated on the character list page that submits a get request to the character view page with the character ID. ID is also refactored everywhere in the code to be a long.

+++ src/main/java/com/connorwiltsie/fallout/charactertracker/entity/PlayerCharacter.java
@@ -12,10 +12,10 @@ public class PlayerCharacter {
     @Column(name = "id")
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     //Associated Player
-    private int playerID;
+    private long playerID;
     
     //General character information
     private String name;
